src/requirements.txt
asgiref==3.6.0
cffi==1.15.1
contourpy==1.0.6
cryptography==39.0.2
cycler==0.11.0
Django==4.1.5
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
fonttools==4.38.0
imbalanced-learn==0.10.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.1
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
mysqlclient==2.2.1
